Language detection part of date_parser moved to language package
